pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "groq",
    "openai",
    "pydantic",
    "pydantic-settings",
    "streamlit",
    "google-genai",
    "pandas>=2.3.0",
    "seaborn>=0.13.2",
    "matplotlib>=3.10.3",
    "qdrant-client>=1.14.3",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-community>=0.3.27",
    "langgraph>=0.5.1",
    "langchain-anthropic>=0.3.17",
    "langchain-openai>=0.3.27",
    "opik>=1.7.42",
]
